pyproject.toml
[tool.poetry]
name = "test-project-for-company"
version = "0.1.0"
description = ""
authors = ["natal <natal@CLOWN>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
pydantic = {extras = ["email"], version = "^2.8.2"}
dynaconf = "^3.2.6"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
